+++ CHANGELOG.md
@@ -10,6 +10,9 @@
   to set HTML attributes and a CSS class for filter cells (@vjik)
 - New #362: Add `GridView::captionAttributes()` method and `$attributes` parameter to `GridView::caption()` to set
   HTML attributes for the `caption` tag (@vjik)
+- New #367: Add `OffsetPagination::currentLinkAttributes()` and `disabledLinkAttributes()`, and
+  `KeysetPagination::disabledLinkAttributes()` to set HTML attributes for the current page link or a disabled
+  control, layered on top of `linkAttributes()` (@vjik)
 - Enh #358: Make dependency container in `GridView` constructor optional (@vjik)
 - Bug #363: Render disabled `KeysetPagination` controls ("previous" on the first page, "next" on the last page) as
   `span` elements instead of `a` elements without an `href` (@vjik)